Move functions from fitting script to stats.py
